[api] refactor project.update_from_xml
